requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain_groq
pymongo